Proposing several editorial changes to README.MD
There are a few spelling and grammatical errors in the main repository README files. Here, I have proposed some changes to improve readability and to better support any readers' efforts to use Google Translate to understand the text.

I have also proposed a few rephrasings and elaborations of instructions to facilitate users who prefer to scan and jump-around in the instructions , rather than reading from top-to-bottom.